Initial commit from files found deep in my hard drive:
ITEMS.DAT appears to be corrupted.  Also, the Free Pascal Compiler
that I found seems to work different from Turbo Pascal on DOS or Win95,
in that the code errors when attempted to reopen an opened file.  The
code will need to be reworked to never open a file more than once, even
for reading.