Global roomfile held open whole time, to avoid reentrance problem:
Instead of having a global position within the room file, so that
any lookups in the file have to attempt to re-open a second copy
of the file; we keep a global 'current room' variable 'r1'.  Any
lookups in files use local variables to store room info, so that
r1 is not changed.  No need to be shifting backward after reading
a room everywhere.

Note that this isn't tested yet, as the corrupted item.dat file
doesn't let the program progress beyond the first room yet.