Feature: poisoning:
Death takes you to a room attached to your own stomach.  The stomach contains
everything you've eaten (including the poison.)  It is connected back to the
starting room.  You can eat things while in your stomach -- it's just a normal
room -- and if you eat something poisonous it will just take you right back to
the death room as normal.